requirements.txt
boto3>=1.26.0
slack-sdk>=3.19.0
requests>=2.28.0
python-dateutil>=2.8.0